odebrani nefunkniho testu
